winedeps: Add Samba to Optional.
Samba provides ntlm_auth and winbindd, which Wine optionally depends on. This doesn't seem to be documented on WineHQ but Wine stdout says those components are missing and to expect problems, whatever they may be. I haven't ever ran into an issue because of that but let's include the package in Optional anyway.