Revert "GLib2: Promote pcre2 to required."
This reverts commit ebb31de16fdd878c90b91a9611291550911a7bde.

The general rule for *LFS is if a package is needed as a dependency for
another package, and can either be compiled for the system beforehand,
or downloaded/packaged as a subproject and compiled then, then it is a
recommended dependency, NOT a required dependency. As such, pcre2 shall
remain recommended and not required. Besides, by this point, stuff like
cURL should be installed already and the user should have a working
internet connection.