Mesa: Comment out patch.
The NVK driver is most likely not structured correctly, so working
around the issue is not recommended. The issue is more complicated than
just Zink falsely reporting the driver setup as false instead of true.
Waiting for upstream to see what they can do about the issue.

If you wish to apply the patch, refer to GLFS releases.