x7lib: Skip over xtrans for lib32.
Previously, the compilation of lib32-xtrans was to just generate a lib32
pkgconf file and for it to be put in /usr/lib32/pkgconfig. Nothing else.
However, this doesn't seem necessary as xtrans doesn't supply any
libraries, and pkgconfig with a PKG_CONFIG_PATH set will ALSO look in
/usr/share/pkgconfig, where xtrans installed it's pkgconf file by
default. Thus, it makes sense to just skip over xtrans in the lib32
build and just use the pkgconf file provided by the normal installation.