NVIDIA: Archived.
I have been angry from this driver for a while now. It honestly used to
be fine. Not anymore. The writing has been on the wall. So it's time for
this driver to get on the chopping block. I will be reviewing future
NVIDIA releases to see if things have changed. But I am not
greenlighting an update from a few days ago that requires a kernel
version older than a month.

The book links to an earlier version for installation of those who want
it.