FFmpeg: Some -> most.
Most dependencies are auto-detected and enabled if detected, for
example, PulseAudio and Vulkan. I do not know if the note is outdated or
not because it's been fine for me and I rarely need to pass much options
to FFmpeg. An outlier is support for NVENC/NVDEC which needs to be
explicitly enabled, but as we do not support CUDA yet, this does not
matter for GLFS.