Internal: Removed PDF and nochunks capability.
This is done for multiple reasons, and most of them stem from the fact
that GLFS does things in a way that struggles to reconcile with the pdf
and nochunks targets. For example, CSS and images is a massive sore
spot. Since they don't get maintained anymore, dragging them behind the
barn and putting them down. Apologies to those that use these targets.