NVIDIA: Improve nvidia_symlink.
Credit to Toxikuu for creating the base of the improvement. I did make a
fair amount of changes though to better handle upgrading and
downgrading. Error handling has been slightly improved. But Bash makes
things pretty difficult to do it properly.

Also, we really need ln -sfv because of soname shenanigans. We really
need to make sure the links get relinked to the new current libraries.
Without proper checking and ln -f, the links would point to old
versions. This isn't a problem for a first install, but absolutely is a
problem on a second install.