NVIDIA: Just have one consistent script.
We now need to stick with this. I think this is truly the most
user-friendly approach while giving a decent amount of flexibility to
users. They can now mix and match via options. With manifest versions,
we can update them if NVIDIA adds more libraries or takes them away. It
also seems like most symlinks aren't necessary. Since optional symlinks
would take quite a while to implement, this does the trick. The install
script is also now POSIX compliant and can be ran with Dash. The
uninstall script still needs to be adapated for POSIX, but at the very
least, we have a Bash shebang for that script.

Why not create the install script manually in the book? It's pretty big
and would cause the NVIDIA page to balloon and it should really be as
short as possible. Anyhow, I'm sticking with this. I'm going to do a
rebuild soon so I'll found out pretty soon if it is borked or not. I did
run it on a most clean system. There was still some old NVIDIA cruft
lying around though, but find didn't reveal much. Anyhow, it seemed to
work just fine. Plasma 6 X11 and Wayland worked, GLX+EGL works on both
instances. Wayland Plasma still is pretty slow unfortunately and I don't
know why. 165 fps on X11 drops to not much higher than 60 which is a
huge drop-off.

And a notice: The rebuild I mentioned will be for my main system and I
will rely on the NVIDIA driver from now on, so NVIDIA updates will be
handled by me unless someone like Tox wants to step in. Though I got it
covered unless NVIDIA pulls another magic trick and bricks something
related to the kernel. I'm just glad they don't seem to brick userspace
like Mesa is doing with Zink+NVK.